fix: fix publishing gradle script to use only one staging repository
