+++ .github/workflows/gradle-publish.yml
@@ -22,5 +22,8 @@ jobs:
           java-version: 8
           distribution: adopt
 
-      - name: Publish to maven central repository
+      - name: Create staging repository and Upload archives to sonatype nexus
         run: gradle publish
+
+      - name: Close staging repository and Publish to maven central repository
+        run: gradle closeAndReleaseRepository
